REFACTORING: Cleaned up API of pre::json::jsonize(yourtype) /  pre::json::dejsonize(yourtype).
