+++ test/dejsonize_test.cpp
@@ -1,26 +1,22 @@
-#define BOOST_TEST_MODULE adapted_struct_jsonize_test
+#define BOOST_TEST_MODULE dejsonize_test
 #include <boost/test/included/unit_test.hpp>
 
 #include <iostream>
-
 #include <vector>
 #include <list>
 
 #include <boost/fusion/include/equal_to.hpp>
-
 #include <boost/fusion/include/define_struct.hpp>
 #include <boost/fusion/include/adapt_struct.hpp>
-#include <pre/boost/fusion/adapted_struct_jsonize.hpp>
-#include <boost/variant.hpp>
 
+#include <boost/variant.hpp>
 #include <boost/tuple/tuple.hpp>
+
 #include <boost/exception_ptr.hpp>
 
-#include <boost/variant.hpp>
-
-
 #include <pre/json/traits/is_jsonizable.hpp>
-
+#include <pre/json/jsonize.hpp>
+#include <pre/json/dejsonize.hpp>
 
 namespace datamodel {
 
@@ -73,7 +69,7 @@ BOOST_FUSION_ADAPT_STRUCT(datamodel::sales_assitant,
   salary,
   main_customer)
 
-BOOST_AUTO_TEST_CASE (adapted_struct_jsonize_test_traits) {
+BOOST_AUTO_TEST_CASE (traits) {
 
   struct non_adapted_struct {
     int dummy;
@@ -95,7 +91,7 @@ BOOST_AUTO_TEST_CASE (adapted_struct_jsonize_test_traits) {
 
 }
 
-BOOST_AUTO_TEST_CASE (adapted_struct_jsonize_test_composedtype) {
+BOOST_AUTO_TEST_CASE (composedtype) {
 
 
   // Testing a customer structure
@@ -106,13 +102,13 @@ BOOST_AUTO_TEST_CASE (adapted_struct_jsonize_test_composedtype) {
     {datamodel::skill{"C++", 10}, datamodel::skill{"GML", 20}, datamodel::skill{"Linux", 2}}
   };
 
-  auto json_customer = boost::fusion::adapted_struct_jsonize::jsonize(customer);
+  auto json_customer = pre::json::jsonize(customer);
   std::cout << json_customer << std::endl;
 
-  auto customer_parsed = boost::fusion::adapted_struct_dejsonize::dejsonize<datamodel::customer>(json_customer);
+  auto customer_parsed = pre::json::dejsonize<datamodel::customer>(json_customer);
   BOOST_ASSERT(customer_parsed == customer);
 
-  auto json_customer_after_reserialize = boost::fusion::adapted_struct_jsonize::jsonize(customer);
+  auto json_customer_after_reserialize = pre::json::jsonize(customer);
   std::cout << json_customer_after_reserialize << std::endl;
   BOOST_ASSERT(json_customer == json_customer_after_reserialize);
 
@@ -122,17 +118,17 @@ BOOST_AUTO_TEST_CASE (adapted_struct_jsonize_test_composedtype) {
     {"experience", 10}
   };
 
-  auto skill_parsed = boost::fusion::adapted_struct_dejsonize::dejsonize<datamodel::skill>(obj);
+  auto skill_parsed = pre::json::dejsonize<datamodel::skill>(obj);
 
   BOOST_ASSERT(skill_parsed.skill_name == "GML");
   BOOST_ASSERT(skill_parsed.experience == 10);
-  std::cout << boost::fusion::adapted_struct_jsonize::jsonize(skill_parsed) << std::endl;
+  std::cout << pre::json::jsonize(skill_parsed) << std::endl;
 }
 
 /**
  * Test nested objects
  */
-BOOST_AUTO_TEST_CASE (adapted_struct_jsonize_test_nested) {
+BOOST_AUTO_TEST_CASE (nested) {
   datamodel::customer customer {
     "Mr. Dupond",
     43,
@@ -141,18 +137,18 @@ BOOST_AUTO_TEST_CASE (adapted_struct_jsonize_test_nested) {
   };
 
   datamodel::sales_assitant assistant{"Mr. Gold", 130000, customer};
-  auto json_sales_assistant = boost::fusion::adapted_struct_jsonize::jsonize(assistant);
+  auto json_sales_assistant = pre::json::jsonize(assistant);
   std::cout << json_sales_assistant << std::endl;
 
-  auto deser_assistant = boost::fusion::adapted_struct_dejsonize::dejsonize<datamodel::sales_assitant>(json_sales_assistant);
+  auto deser_assistant = pre::json::dejsonize<datamodel::sales_assitant>(json_sales_assistant);
 
-  auto json_sales_assistant_reserialized = boost::fusion::adapted_struct_jsonize::jsonize(deser_assistant);
+  auto json_sales_assistant_reserialized = pre::json::jsonize(deser_assistant);
   std::cout << json_sales_assistant_reserialized << std::endl;
 
   BOOST_ASSERT(json_sales_assistant_reserialized == json_sales_assistant);
 }
 
-BOOST_AUTO_TEST_CASE (adapted_struct_jsonize_test_working_plain_json) {
+BOOST_AUTO_TEST_CASE (working_plain_json) {
   auto json = R"(
     {
         "main_customer": {
@@ -183,13 +179,13 @@ BOOST_AUTO_TEST_CASE (adapted_struct_jsonize_test_working_plain_json) {
 
   )"_json;
 
-  auto deser_assistant = boost::fusion::adapted_struct_dejsonize::dejsonize<datamodel::sales_assitant>(json);
+  auto deser_assistant = pre::json::dejsonize<datamodel::sales_assitant>(json);
 
-  auto json_sales_assistant_reserialized = boost::fusion::adapted_struct_jsonize::jsonize(deser_assistant);
+  auto json_sales_assistant_reserialized = pre::json::jsonize(deser_assistant);
   std::cout << json_sales_assistant_reserialized << std::endl;
 }
 
-BOOST_AUTO_TEST_CASE (adapted_struct_jsonize_test_incorrect_plain_json) {
+BOOST_AUTO_TEST_CASE (incorrect_plain_json) {
   // Test buggy keys
   try {
     auto json = R"(
@@ -222,7 +218,7 @@ BOOST_AUTO_TEST_CASE (adapted_struct_jsonize_test_incorrect_plain_json) {
 
     )"_json;
 
-    auto deser_assistant = boost::fusion::adapted_struct_dejsonize::dejsonize<datamodel::sales_assitant>(json);
+    auto deser_assistant = pre::json::dejsonize<datamodel::sales_assitant>(json);
     // Must fail
     BOOST_ASSERT(false);
 
@@ -264,7 +260,7 @@ BOOST_AUTO_TEST_CASE (adapted_struct_jsonize_test_incorrect_plain_json) {
 
     )"_json;
 
-    auto deser_assistant = boost::fusion::adapted_struct_dejsonize::dejsonize<datamodel::sales_assitant>(json);
+    auto deser_assistant = pre::json::dejsonize<datamodel::sales_assitant>(json);
     // Must fail
     BOOST_ASSERT(false);
 
@@ -276,7 +272,7 @@ BOOST_AUTO_TEST_CASE (adapted_struct_jsonize_test_incorrect_plain_json) {
 }
 
 
-BOOST_AUTO_TEST_CASE (adapted_struct_jsonize_test_containers_direct) {
+BOOST_AUTO_TEST_CASE (containers_direct) {
   std::vector<datamodel::skill> skills {
     datamodel::skill{"C++", 0},
     datamodel::skill{"Java", 0},
@@ -287,10 +283,10 @@ BOOST_AUTO_TEST_CASE (adapted_struct_jsonize_test_containers_direct) {
     datamodel::skill{"Project-Management", 0}
   };
 
-  auto skills_json = boost::fusion::adapted_struct_jsonize::jsonize(skills);
+  auto skills_json = pre::json::jsonize(skills);
   std::cout << skills_json << std::endl;
 
-  auto skills_deserialized = boost::fusion::adapted_struct_dejsonize::dejsonize<std::vector<datamodel::skill>>(skills_json); 
+  auto skills_deserialized = pre::json::dejsonize<std::vector<datamodel::skill>>(skills_json); 
   BOOST_ASSERT(skills == skills_deserialized);
 }
 
@@ -347,7 +343,7 @@ BOOST_FUSION_ADAPT_STRUCT(datamodel::employee,
   name,
   responsibility)
 
-BOOST_AUTO_TEST_CASE (adapted_struct_jsonize_test_vairants) {
+BOOST_AUTO_TEST_CASE (vairants) {
 
   std::vector<datamodel::employee> employees {
     {"King", datamodel::cashier{"hardware", 1} },
@@ -356,12 +352,12 @@ BOOST_AUTO_TEST_CASE (adapted_struct_jsonize_test_vairants) {
     {"Ward", datamodel::cashier{"Food", 2} }
   };
 
-  auto employees_json = boost::fusion::adapted_struct_jsonize::jsonize(employees);
+  auto employees_json = pre::json::jsonize(employees);
   std::cout << employees_json << std::endl;
 
-  auto employees_deserialized = boost::fusion::adapted_struct_dejsonize::dejsonize<std::vector<datamodel::employee>>(employees_json); 
+  auto employees_deserialized = pre::json::dejsonize<std::vector<datamodel::employee>>(employees_json); 
 
-  auto employees_reserialized = boost::fusion::adapted_struct_jsonize::jsonize(employees_deserialized);
+  auto employees_reserialized = pre::json::jsonize(employees_deserialized);
   std::cout << employees_reserialized << std::endl;
 
   BOOST_ASSERT(employees == employees_deserialized);
@@ -378,7 +374,7 @@ namespace datamodel {
 
 BOOST_FUSION_ADAPT_STRUCT(datamodel::struct_with_a_map, with_a_map, some_other_map)
 
-BOOST_AUTO_TEST_CASE(adapted_struct_jsonize_test_maps) {
+BOOST_AUTO_TEST_CASE(maps) {
 
   datamodel::struct_with_a_map some{
     {
@@ -390,10 +386,10 @@ BOOST_AUTO_TEST_CASE(adapted_struct_jsonize_test_maps) {
     }
   };
 
-  auto some_json = boost::fusion::adapted_struct_jsonize::jsonize(some);
+  auto some_json = pre::json::jsonize(some);
   std::cout << some_json << std::endl; 
 
-  auto some_deserialized = boost::fusion::adapted_struct_dejsonize::dejsonize<datamodel::struct_with_a_map>(some_json);
+  auto some_deserialized = pre::json::dejsonize<datamodel::struct_with_a_map>(some_json);
   
   BOOST_ASSERT(some_deserialized == some); 
 }
@@ -425,16 +421,16 @@ BOOST_FUSION_ADAPT_STRUCT(datamodel::some_value,
   value,
   enums_without_underlying)
 
-BOOST_AUTO_TEST_CASE (adapted_struct_jsonize_test_enums) {
+BOOST_AUTO_TEST_CASE (enums) {
 
   datamodel::some_value val{datamodel::my_value_type::TYPE2, 12, datamodel::THREE};
 
-  auto val_json = boost::fusion::adapted_struct_jsonize::jsonize(val);
+  auto val_json = pre::json::jsonize(val);
   std::cout << val_json << std::endl;
 
-  auto val_deserialized = boost::fusion::adapted_struct_dejsonize::dejsonize<datamodel::some_value>(val_json); 
+  auto val_deserialized = pre::json::dejsonize<datamodel::some_value>(val_json); 
 
-  auto val_reserialized = boost::fusion::adapted_struct_jsonize::jsonize(val_deserialized);
+  auto val_reserialized = pre::json::jsonize(val_deserialized);
   std::cout << val_reserialized << std::endl;
 
   BOOST_ASSERT(val == val_deserialized);
@@ -444,12 +440,12 @@ BOOST_AUTO_TEST_CASE (adapted_struct_jsonize_chrono) {
 
   boost::chrono::milliseconds val{100};
 
-  auto val_json = boost::fusion::adapted_struct_jsonize::jsonize(val);
+  auto val_json = pre::json::jsonize(val);
   std::cout << val_json << std::endl;
 
-  auto val_deserialized = boost::fusion::adapted_struct_dejsonize::dejsonize<boost::chrono::milliseconds>(val_json); 
+  auto val_deserialized = pre::json::dejsonize<boost::chrono::milliseconds>(val_json); 
 
-  auto val_reserialized = boost::fusion::adapted_struct_jsonize::jsonize(val_deserialized);
+  auto val_reserialized = pre::json::jsonize(val_deserialized);
   std::cout << val_reserialized << std::endl;
 
   BOOST_ASSERT(val == val_deserialized);
@@ -473,12 +469,12 @@ BOOST_AUTO_TEST_CASE (adapted_struct_jsonize_chrono_composed) {
 
   datamodel::some_time val{"my time", boost::chrono::milliseconds(100), std::chrono::minutes(43)};
 
-  auto val_json = boost::fusion::adapted_struct_jsonize::jsonize(val);
+  auto val_json = pre::json::jsonize(val);
   std::cout << val_json << std::endl;
 
-  auto val_deserialized = boost::fusion::adapted_struct_dejsonize::dejsonize<datamodel::some_time>(val_json); 
+  auto val_deserialized = pre::json::dejsonize<datamodel::some_time>(val_json); 
 
-  auto val_reserialized = boost::fusion::adapted_struct_jsonize::jsonize(val_deserialized);
+  auto val_reserialized = pre::json::jsonize(val_deserialized);
   std::cout << val_reserialized << std::endl;
 
   BOOST_ASSERT(val == val_deserialized);
@@ -495,34 +491,34 @@ BOOST_FUSION_ADAPT_STRUCT(datamodel::with_optional_member,
   field_optional,
   other)
 
-BOOST_AUTO_TEST_CASE (adapted_struct_jsonize_optional) {
+BOOST_AUTO_TEST_CASE (optional) {
 
   boost::optional<std::string> val{"some text"};
 
-  auto val_json = boost::fusion::adapted_struct_jsonize::jsonize(val);
+  auto val_json = pre::json::jsonize(val);
   std::cout << val_json << std::endl;
 
-  auto val_deserialized = boost::fusion::adapted_struct_dejsonize::dejsonize<boost::optional<std::string>>(val_json); 
+  auto val_deserialized = pre::json::dejsonize<boost::optional<std::string>>(val_json); 
 
-  auto val_reserialized = boost::fusion::adapted_struct_jsonize::jsonize(val_deserialized);
+  auto val_reserialized = pre::json::jsonize(val_deserialized);
   std::cout << val_reserialized << std::endl;
 
   BOOST_ASSERT(val == val_deserialized);
 }
 
-BOOST_AUTO_TEST_CASE (adapted_struct_jsonize_optional_member) {
+BOOST_AUTO_TEST_CASE (optional_member) {
 
   {
     datamodel::with_optional_member val;
     val.field_optional = 12.32;
     val.other = 13.21;
 
-    auto val_json = boost::fusion::adapted_struct_jsonize::jsonize(val);
+    auto val_json = pre::json::jsonize(val);
     std::cout << val_json << std::endl;
 
-    auto val_deserialized = boost::fusion::adapted_struct_dejsonize::dejsonize<datamodel::with_optional_member>(val_json); 
+    auto val_deserialized = pre::json::dejsonize<datamodel::with_optional_member>(val_json); 
 
-    auto val_reserialized = boost::fusion::adapted_struct_jsonize::jsonize(val_deserialized);
+    auto val_reserialized = pre::json::jsonize(val_deserialized);
     std::cout << val_reserialized << std::endl;
 
     BOOST_ASSERT(val == val_deserialized);
@@ -532,12 +528,12 @@ BOOST_AUTO_TEST_CASE (adapted_struct_jsonize_optional_member) {
     datamodel::with_optional_member val;
     val.other = 43.21;
 
-    auto val_json = boost::fusion::adapted_struct_jsonize::jsonize(val);
+    auto val_json = pre::json::jsonize(val);
     std::cout << val_json << std::endl;
 
-    auto val_deserialized = boost::fusion::adapted_struct_dejsonize::dejsonize<datamodel::with_optional_member>(val_json); 
+    auto val_deserialized = pre::json::dejsonize<datamodel::with_optional_member>(val_json); 
 
-    auto val_reserialized = boost::fusion::adapted_struct_jsonize::jsonize(val_deserialized);
+    auto val_reserialized = pre::json::jsonize(val_deserialized);
     std::cout << val_reserialized << std::endl;
 
     BOOST_ASSERT(val == val_deserialized);
